# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ecom API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/admin/brands": {
            "post": {
                "description": "Create a new brand with the specified name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brands"
                ],
                "summary": "Create a new brand.",
                "parameters": [
                    {
                        "description": "Brand name to be created",
                        "name": "brand",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.brands"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created brand",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/success": {
            "get": {
                "description": "Displays a success page after payment completion.",
                "produces": [
                    "text/html"
                ],
                "summary": "Payment Success",
                "operationId": "Success",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment success page",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Error while displaying success page",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.brands": {
            "type": "object",
            "properties": {
                "brandName": {
                    "type": "string"
                }
            }
        },
...
    }
}